requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
discord.py==2.4.0
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
magic_profanity==1.1.1
multidict==6.1.0
mutagen==1.47.0
openai==0.28.0
propcache==0.2.0
psycopg2==2.9.10
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.2.0
sniffio==1.3.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.15.5
youtube-dl==2021.12.17
yt-dlp==2024.10.7
